TimePeriodFormatter as a separate chunk. Fixed bug that crashed when period was longer than an int.

+++ src/main/scala/metl/model/ServiceCheckConfigurator.scala
@@ -11,7 +11,7 @@ object ServiceCheckConfigurator extends ConfigFileReader with Logger {
     trace("loading xml: %s".format(xml))
     (xml \\ "serviceCheck").map(sc => {
       debug("Configuring serviceCheck from:\n%s".format(sc))
-      val periodInt = getInt(sc,"period").getOrElse(60) * 1000
+      val periodInt = (getLong(sc,"period").getOrElse(60L)) * 1000
       val period = new TimeSpan(periodInt)
       val serviceCheckName = getAttr(sc,"name").getOrElse("unnamed")
       val label = getText(sc,"label").getOrElse("no label")
